+++ example/websocket/server/coro-ssl/websocket_server_coro_ssl.cpp
@@ -66,6 +66,15 @@ do_session(
     if(ec)
         return fail(ec, "handshake");
 
+    // Turn off the timeout on the tcp_stream, because
+    // the websocket stream has its own timeout system.
+    beast::get_lowest_layer(ws).expires_never();
+
+    // Set suggested timeout settings for the websocket
+    ws.set_option(
+        websocket::stream_base::suggested_settings(
+            websocket::role_type::server));
+
     // Accept the websocket handshake
     ws.async_accept(yield[ec]);
     if(ec)