+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 216:
 
 * Refactor websocket::stream operations
 * Add websocket::stream timeouts
+* Use suggested timeouts in Websocket examples
 
 --------------------------------------------------------------------------------